# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  city?: string | null;
  isPremium: boolean;
  /** @nullable */
  premiumExpiresAt?: string | null;
  resumeCount?: number;
  coverLetterCount?: number;
  interviewCount?: number;
  createdAt: string;
}

export interface UserRegistration {
  name: string;
  email: string;
  password: string;
  mobile?: string;
  city?: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface UserWithToken {
  user: User;
  token: string;
}

export interface WeeklyActivityItem {
  day: string;
  count: number;
}

export interface Resume {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  projects?: string | null;
  /** @nullable */
  workExperience?: string | null;
  /** @nullable */
  certifications?: string | null;
  /** @nullable */
  languages?: string | null;
  template: string;
  /** @nullable */
  generatedContent?: string | null;
  /** @nullable */
  atsScore?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface DashboardStats {
  resumeCount: number;
  interviewCount: number;
  coverLetterCount: number;
  profileStrength: number;
  isPremium?: boolean;
  weeklyActivity: WeeklyActivityItem[];
  recentResumes?: Resume[];
}

export interface ResumeInput {
  title: string;
  fullName?: string;
  mobile?: string;
  email?: string;
  city?: string;
  education?: string;
  college?: string;
  skills?: string;
  projects?: string;
  workExperience?: string;
  certifications?: string;
  languages?: string;
  template: string;
}

export interface ResumeUpdate {
  title?: string;
  fullName?: string;
  mobile?: string;
  email?: string;
  city?: string;
  education?: string;
  college?: string;
  skills?: string;
  projects?: string;
  workExperience?: string;
  certifications?: string;
  languages?: string;
  template?: string;
  generatedContent?: string;
  atsScore?: number;
}

export interface ResumeScore {
  overallScore: number;
  atsScore: number;
  skillsScore: number;
  formattingScore: number;
  suggestions: string[];
  strengths?: string[];
}

export interface ResumeAnalysisInput {
  resumeText: string;
  jobRole?: string;
}

export interface JobMatchInput {
  resumeText: string;
  jobDescription: string;
}

export interface JobMatchResult {
  matchPercentage: number;
  missingSkills: string[];
  matchedSkills?: string[];
  suggestions: string[];
}

export interface CoverLetter {
  id: number;
  userId: number;
  jobRole: string;
  companyName: string;
  experienceLevel: string;
  content: string;
  createdAt: string;
}

export interface CoverLetterInput {
  jobRole: string;
  companyName: string;
  experienceLevel: string;
  additionalInfo?: string;
}

export interface InterviewSession {
  id: number;
  userId: number;
  category: string;
  questionCount: number;
  /** @nullable */
  avgCommunicationScore?: number | null;
  /** @nullable */
  avgConfidenceScore?: number | null;
  createdAt: string;
}

export interface InterviewSessionInput {
  category: string;
}

export interface InterviewAnswerInput {
  sessionId: number;
  question: string;
  answer: string;
  category?: string;
}

export interface InterviewFeedback {
  feedback: string;
  communicationScore: number;
  confidenceScore: number;
  improvedAnswer: string;
  strengths?: string[];
  improvements?: string[];
}

export interface EnglishImprovementInput {
  text: string;
  context?: string;
}

export interface EnglishImprovementResult {
  improvedText: string;
  corrections: string[];
  professionalTips: string[];
}

export interface Payment {
  id: number;
  /** @nullable */
  userId?: number | null;
  fullName: string;
  mobile: string;
  upiTransactionId: string;
  amount: number;
  plan: string;
  status: string;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
}

export interface PaymentInput {
  fullName: string;
  mobile: string;
  upiTransactionId: string;
  amount: number;
  plan: string;
}

export interface QrCode {
  imageUrl: string;
  type: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface QrCodeInput {
  imageUrl: string;
  type: string;
}

export interface Supporter {
  id: number;
  name: string;
  mobile: string;
  amount: number;
  upiTransactionId: string;
  createdAt: string;
}

export interface SupporterInput {
  name: string;
  mobile: string;
  amount: number;
  upiTransactionId: string;
}

export interface PublicStats {
  resumesCreated: number;
  interviewsPracticed: number;
  activeUsers: number;
  successStories: number;
}

export interface AdminStats {
  totalUsers: number;
  totalResumes: number;
  totalInterviews: number;
  totalCoverLetters: number;
  totalRevenue: number;
  pendingPayments: number;
  totalSupporters: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  isPremium: boolean;
  /** @nullable */
  premiumExpiresAt?: string | null;
  resumeCount: number;
  interviewCount: number;
  coverLetterCount?: number;
  createdAt: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminSession {
  token: string;
  role: string;
}