# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Career Boost AI India API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cover-letters:
    get:
      operationId: listCoverLetters
      tags: [cover-letters]
      summary: List user cover letters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoverLetter"
    post:
      operationId: createCoverLetter
      tags: [cover-letters]
      summary: Generate a cover letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetter"
  /cover-letters/{id}:
    get:
      operationId: getCoverLetter
      tags: [cover-letters]
      summary: Get a cover letter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetter"
    delete:
      operationId: deleteCoverLetter
      tags: [cover-letters]
      summary: Delete a cover letter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Interview ────────────────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/payments/{id}/approve:
    post:
      operationId: adminApprovePayment
      tags: [admin]
      summary: Admin — approve a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/payments/{id}/reject:
    post:
      operationId: adminRejectPayment
      tags: [admin]
      summary: Admin — reject a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/qr:
    put:
      operationId: adminUpdateQr
      tags: [admin]
      summary: Admin — update QR code URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrCodeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── User schemas ─────────────────────────────────────────────────────────
...
    CoverLetter:
      type: object
      required: [id, userId, jobRole, companyName, experienceLevel, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        jobRole:
          type: string
        companyName:
          type: string
        experienceLevel:
          type: string
        content:
          type: string
        createdAt:
          type: string
    CoverLetterInput:
      type: object
      required: [jobRole, companyName, experienceLevel]
      properties:
        jobRole:
          type: string
        companyName:
          type: string
        experienceLevel:
          type: string
        additionalInfo:
          type: string

    # ─── Interview schemas ────────────────────────────────────────────────────
...
    Payment:
      type: object
      required: [id, fullName, mobile, upiTransactionId, amount, plan, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        fullName:
          type: string
        mobile:
          type: string
        upiTransactionId:
          type: string
        amount:
          type: integer
        plan:
          type: string
        status:
          type: string
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
...
    QrCode:
      type: object
      required: [imageUrl, type]
      properties:
        imageUrl:
          type: string
        type:
          type: string
        updatedAt:
          type: ["string", "null"]
    QrCodeInput:
      type: object
      required: [imageUrl, type]
      properties:
        imageUrl:
          type: string
        type:
          type: string

    # ─── Supporter schemas ────────────────────────────────────────────────────
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string