# lib/api-zod/src/generated/types/coverLetter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Boost AI India API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetter {
  id: number;
  userId: number;
  jobRole: string;
  companyName: string;
  experienceLevel: string;
  content: string;
  createdAt: string;
}